Revert "refactor вынес функцию проверки запроса"
This reverts commit 71edc4c9bfc19f924a675497012745f214df9a7e.

+++ controllers/cardsControllers.js
@@ -8,21 +8,14 @@ function errorHandler(error, res) {
         .join(', ')}`,
     });
   }
-  return res.status(500).send({ message: 'Server Error' });
-}
 
-async function checkCorrectRequest(req) {
-  if (req.params.cardId.length <= 20) {
-    return { status: 404, message: { message: 'Некорректный id' } };
+  if (error.name === 'CastError') {
+    return res.status(404).send({
+      message: 'Запрашиваемая карточка не найдена',
+    });
   }
-  const cardForChange = await Card.findById(req.params.cardId);
-  if (cardForChange == null) {
-    return {
-      status: 404,
-      message: { message: 'Запрашиваемая карточка не найдена' },
-    };
-  }
-  return [];
+
+  return res.status(500).send({ message: 'Server Error' });
 }
 
 export const getAllCards = async (req, res) => {
@@ -47,11 +40,15 @@ export const addNewCard = async (req, res) => {
 
 export const deleteCardById = async (req, res) => {
   try {
-    const errCheck = await checkCorrectRequest(req);
-    if (errCheck !== []) {
-      res.status(errCheck.status).send(errCheck.message);
+    if (req.params.cardId.length <= 20) {
+      res.status(400).send({ message: 'Некорректный id' });
+      return;
     }
-
+    const cardForChange = await Card.findById(req.params.cardId);
+    if (cardForChange == null) {
+      res.status(404).send({ message: 'Запрашиваемая карточка не найдена' });
+      return;
+    }
     await Card.findByIdAndDelete(req.params.cardId);
     res.status(200).send({ message: 'Успешно удалено!' });
   } catch (error) {
@@ -61,9 +58,14 @@ export const deleteCardById = async (req, res) => {
 
 export const likeCard = async (req, res) => {
   try {
-    const errCheck = await checkCorrectRequest(req);
-    if (errCheck !== []) {
-      res.status(errCheck.status).send(errCheck.message);
+    if (req.params.cardId.length <= 20) {
+      res.status(400).send({ message: 'Некорректный id' });
+      return;
+    }
+    const cardForChange = await Card.findById(req.params.cardId);
+    if (cardForChange == null) {
+      res.status(404).send({ message: 'Запрашиваемая карточка не найдена' });
+      return;
     }
 
     const card = await Card.findByIdAndUpdate(
@@ -79,11 +81,15 @@ export const likeCard = async (req, res) => {
 
 export const dislikeCard = async (req, res) => {
   try {
-    const errCheck = await checkCorrectRequest(req);
-    if (errCheck !== []) {
-      res.status(errCheck.status).send(errCheck.message);
+    if (req.params.cardId.length <= 20) {
+      res.status(400).send({ message: 'Некорректный id' });
+      return;
     }
-
+    const cardForChange = await Card.findById(req.params.cardId);
+    if (cardForChange == null) {
+      res.status(404).send({ message: 'Запрашиваемая карточка не найдена' });
+      return;
+    }
     const card = await Card.findByIdAndUpdate(
       req.params.cardId,
       { $pull: { likes: req.user._id } }, // убрать _id из массива
